backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.4.6
twilio==9.0.5
anthropic==0.31.0
apscheduler==3.10.4
httpx==0.27.0
pydantic==2.9.0
python-multipart==0.0.9
pytz==2024.1
google-generativeai==0.8.3
